@@ -10,7 +10,8 @@
   "scripts": {
     "start": "node lib/index.js",
     "build": "rm -rf lib && tsc",
-    "lint": "tslint -c tslint.json 'src/**/*.ts'"
+    "lint": "eslint 'src/**/*.ts'",
+    "test": "mocha"
   },
   "bugs": {
     "url": "https://github.com/ajaysinghj8/osham/issues"
@@ -31,34 +32,44 @@
   },
   "devDependencies": {
     "@types/debug": "^4.1.5",
-    "@types/follow-redirects": "^1.8.0",
-    "@types/js-yaml": "^3.12.2",
+    "@types/follow-redirects": "^1.13.0",
+    "@types/js-yaml": "^4.0.0",
     "@types/koa-compose": "^3.2.5",
-    "@types/node": "^13.7.6",
-    "@types/redis": "^2.8.16",
-    "artillery": "^1.6.0-29",
+    "@types/node": "^14.14.21",
+    "@types/parseurl": "^1.3.1",
+    "@types/path-to-regexp": "^1.7.0",
+    "@types/redis": "^2.8.28",
+    "@types/statuses": "^2.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.13.0",
+    "@typescript-eslint/parser": "^4.13.0",
+    "artillery": "^1.6.1",
+    "eslint": "^7.18.0",
+    "eslint-config-prettier": "^7.1.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-prettier": "^3.3.1",
     "flamebearer": "^1.1.3",
+    "mocha": "^8.2.1",
     "mocha-typescript": "^1.1.17",
-    "supertest": "^4.0.2",
-    "ts-node": "^8.6.2",
-    "tslint": "^6.0.0",
-    "typescript": "^3.8.2"
+    "prettier": "^2.2.1",
+    "supertest": "^6.1.1",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
   },
   "dependencies": {
-    "debug": "^4.1.1",
+    "debug": "^4.3.1",
     "dotenv": "^8.2.0",
-    "follow-redirects": "^1.10.0",
-    "js-yaml": "^3.13.1",
+    "follow-redirects": "^1.13.1",
+    "js-yaml": "^4.0.0",
     "koa-compose": "^4.1.0",
-    "parse-duration": "^0.1.2",
+    "parse-duration": "^0.4.4",
     "parseurl": "^1.3.3",
     "path-to-regexp": "0.1.7",
     "querystring": "^0.2.0",
     "redis": "^3.0.2",
     "reflect-metadata": "^0.1.13",
-    "statuses": "^1.5.0"
+    "statuses": "^2.0.1"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=14.0.0"
   }
 }
