@@ -1,10 +1,11 @@
-import { Context } from '../ctx.provider';
+import * as Koa from 'koa';
+import { IContext } from '../types';
 
-export async function HealthCheck(ctx: Context, next: any) {
-    if (ctx.path === '/health') {
-        ctx.body = 'ok';
-        ctx.statusCode = 200;
-        return;
-    }
-    await next();
+export async function HealthCheck(ctx: IContext, next: Koa.Next): Promise<void> {
+  if (ctx.path === '/health') {
+    ctx.body = 'ok';
+    ctx.statusCode = 200;
+    return;
+  }
+  await next();
 }
