@@ -1,8 +1,9 @@
-import { Context } from '../ctx.provider';
+import * as Koa from 'koa';
+import { IContext } from '../types';
 
-export async function RouteTimeReqRes(ctx: Context, next: any) {
-    const start = +new Date();
-    await next();
-    const ms = +new Date() - start;
-    ctx.set('X-Response-Time', `${ms}ms`);
+export async function RouteTimeReqRes(ctx: IContext, next: Koa.Next): Promise<void> {
+  const start = +new Date();
+  await next();
+  const ms = +new Date() - start;
+  ctx.set('x-osham-time', `${ms}ms`);
 }
