@@ -1,18 +1,19 @@
 import { OutgoingHttpHeaders } from 'http';
-import { Context } from '../ctx.provider';
+import { IContext } from '../types';
 
-export function writeHeaders(headers: OutgoingHttpHeaders, ctx?: Context) {
-    const headersMap: any = {};
-    const setCtxHeaders = (field: string, value: any) => {
-        ctx.set(field, value);
-        headersMap[field] = value;
-    };
-    const setHeader = ctx ? setCtxHeaders : (field: string, value: any) => headersMap[field] = value;
-    for (const field in headers) {
-        if (!Object.prototype.hasOwnProperty.call(headers, field)) continue;
-        const value = headers[field];
-        const val = Array.isArray(value) ? value.map(String) : String(value);
-        setHeader(String(field).trim(), val);
-    }
-    return headersMap;
+export function writeHeaders(headers: OutgoingHttpHeaders, ctx?: IContext): Record<string, string> {
+  const headersMap: Record<string, string> = {};
+  const setCtxHeaders = (field: string, value: string) => {
+    ctx.set(field, value);
+    headersMap[field] = value;
+  };
+  const setHeader = ctx ? setCtxHeaders : (field: string, value: string) => (headersMap[field] = value);
+  for (const field in headers) {
+    if (!Object.prototype.hasOwnProperty.call(headers, field)) continue;
+    const value = headers[field];
+    const val = Array.isArray(value) ? value.map(String) : String(value);
+    if (!val) continue;
+    setHeader(String(field).trim(), val as string);
+  }
+  return headersMap;
 }
