@@ -5,52 +5,51 @@ import { timeOutResponse } from '../utils';
 
 const logger = Debug('acp:service:pool');
 
-const TIMEOUT: number = (+process.env.TIMEOUT) || 5000;
+const TIMEOUT: number = +process.env.TIMEOUT || 5000;
 
 export class RequestPool {
-    static pool: Set<string> = new Set();
-    static ee: EventEmitter = (new EventEmitter()).setMaxListeners(Infinity);
+  static pool: Set<string> = new Set();
+  static ee: EventEmitter = new EventEmitter().setMaxListeners(Infinity);
 
-    static has(key: string) {
-        return RequestPool.pool.has(key);
-    }
-    static add(key: string) {
-        RequestPool.pool.add(key);
-        return RequestPool;
-    }
+  static has(key: string): boolean {
+    return RequestPool.pool.has(key);
+  }
+  static add(key: string): RequestPool {
+    RequestPool.pool.add(key);
+    return RequestPool;
+  }
 
-    static wait(key: string) {
-        logger(`wating for ${key}`);
-        return new Promise((resolve) => {
-
-            const timeout = setTimeout(() => {
-                RequestPool.pool.delete(key);
-                RequestPool.ee.off(key, handler);
-                logger(`TIMEOUT ${key}`);
-                handler(timeOutResponse('RequestPool failed'));
-            }, TIMEOUT);
-
-            function handler(data: any) {
-                clearTimeout(timeout);
-                resolve(data);
-                logger(`respond for ${key}`);
-            }
-
-            RequestPool.ee.once(key, handler);
-        });
-    }
-
-    static async putAndPublish(key: string, data: any, expires: number) {
-        logger(`putting and publishing for ${key}`);
+  static wait(key: string): Promise<unknown> {
+    logger(`wating for ${key}`);
+    return new Promise(resolve => {
+      const timeout = setTimeout(() => {
         RequestPool.pool.delete(key);
-        RequestPool.ee.emit(key, data);
-        await Cache.put(key, data, expires);
-        return data;
-    }
+        RequestPool.ee.off(key, handler);
+        logger(`TIMEOUT ${key}`);
+        handler(timeOutResponse('RequestPool failed'));
+      }, TIMEOUT);
 
-    static async errorAndPublish(key: string, data: any) {
-        logger(`error and publishing for ${key}`);
-        RequestPool.pool.delete(key);
-        RequestPool.ee.emit(key, data);
-    }
+      function handler(data: unknown) {
+        clearTimeout(timeout);
+        resolve(data);
+        logger(`respond for ${key}`);
+      }
+
+      RequestPool.ee.once(key, handler);
+    });
+  }
+
+  static async putAndPublish(key: string, data: unknown, expires: number): Promise<unknown> {
+    logger(`putting and publishing for ${key}`);
+    RequestPool.pool.delete(key);
+    RequestPool.ee.emit(key, data);
+    await Cache.put(key, data, expires);
+    return data;
+  }
+
+  static async errorAndPublish(key: string, data: unknown): Promise<void> {
+    logger(`error and publishing for ${key}`);
+    RequestPool.pool.delete(key);
+    RequestPool.ee.emit(key, data);
+  }
 }
