@@ -1,12 +1,13 @@
-
-import { createClient } from 'redis';
+import { createClient, RedisClient } from 'redis';
 import * as Debug from 'debug';
 
 const logger = Debug('acp:redis');
 
-export const Redis = createClient({
+export const RedisClientProvider = (): RedisClient => {
+  const Connection = createClient({
     host: process.env.REDIS_HOST,
     port: +process.env.REDIS_PORT,
-});
-
-Redis.on('error', (e) => logger(e));
+  });
+  Connection.on('error', e => logger(e));
+  return Connection;
+};
